// src/components/ble-settings/types.ts
export type BLESettingsDrawerPanelGestureHandlers = {
    onPointerDown: PointerEventHandler<HTMLElement>;
    onPointerMove: PointerEventHandler<HTMLElement>;
    onPointerUp: PointerEventHandler<HTMLElement>;
    onPointerCancel: PointerEventHandler<HTMLElement>;
    onLostPointerCapture: PointerEventHandler<HTMLElement>;
};